3.2 Server info route now responds with number of people in phonebook and request time.
