+++ src/server.js
@@ -1,29 +1,8 @@
 const express = require('express');
 const app = express();
 const port = 3001;
+const data = require('./data');
 
-const data = [
-    { 
-      "id": 1,
-      "name": "Arto Hellas", 
-      "number": "040-123456"
-    },
-    { 
-      "id": 2,
-      "name": "Ada Lovelace", 
-      "number": "39-44-5323523"
-    },
-    { 
-      "id": 3,
-      "name": "Dan Abramov", 
-      "number": "12-43-234345"
-    },
-    { 
-      "id": 4,
-      "name": "Mary Poppendieck", 
-      "number": "39-23-6423122"
-    }
-];
 
 app.listen(port, () => {
     console.log(`Listening on port ${port}...`);
@@ -31,4 +10,9 @@ app.listen(port, () => {
 
 app.get('/api/persons', (req, res) => {
     res.send(data);
+})
+
+app.get('/info', (req, res) => {
+    const phonebookEntries = Object.keys(data).length;
+    res.send(`Phonebook has info for ${phonebookEntries} people. \n ${Date()}`);
 })
